requirements.txt
alembic==1.8.1
anyio==3.6.1
asyncpg==0.26.0
atomicwrites==1.4.1
attrs==21.4.0
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
databases==0.6.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
jose==1.0.0
jwt==1.3.1
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.2
